+++ src/lib.rs
@@ -392,10 +392,10 @@ impl Tensor {
 /// - Panics if contracted sizes don't match
 #[must_use]
 pub fn contract(
-    out_indices: &[u32],
-    a_indices: &[u32],
+    out_indices: &[usize],
+    a_indices: &[usize],
     a: &Tensor,
-    b_indices: &[u32],
+    b_indices: &[usize],
     b: &Tensor,
 ) -> Tensor {
     assert_eq!(a_indices.len(), a.ndim());
